Refactor ControlPanel, MainGame, and Settings components to remove repeatTrying prop
- Removed the repeatTrying prop from ControlPanel and Settings components to streamline their interfaces.
- Updated MainGame to utilize new useGamePlay and useBGM hooks for improved game logic and audio management.
- Enhanced game initialization and state handling, improving overall user experience and code maintainability.
